requirements.txt
Django==2.2.7
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-api-key==1.4.0
djangorestframework-filters==0.11.1
psycopg2-binary==2.8.2
django-rest-passwordreset
django-rest-swagger
django-cors-headers
django-datatables-view
Pillow
django-settings-export
django-multiselectfield
django-countries